+++ src/sll/assembly.c
@@ -1034,7 +1034,7 @@ static const sll_node_t* _generate_on_stack(const sll_node_t* o,assembly_generat
 			{
 				sll_arg_count_t l=o->dt.ac;
 				if (!l){
-					GENERATE_OPCODE(g_dt,SLL_ASSEMBLY_INSTRUCTION_TYPE_PUSH_NULL);
+					GENERATE_OPCODE(g_dt,SLL_ASSEMBLY_INSTRUCTION_TYPE_PUSH_ZERO);
 					return o+1;
 				}
 				o=_generate_on_stack(o+1,g_dt);